# demo/src/main/resources/static/doc/beerswagger.yaml
openapi: 3.0.1
info:
  title: Beers
  description: 'This is a sample server Beers server.  You can find out more about
    Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For
    this sample, you can use the api key `special-key` to test the authorization filters.'
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.5
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: http://localhost:8080/api/v1
- url: https://localhost:8080/api/v1
tags:
- name: beer
  description: Everything about your Beers
  externalDocs:
    description: Find out more
    url: http://swagger.io
- name: pub
  description: Everything about your Beers
  externalDocs:
    description: Find out more
    url: http://swagger.io
paths:
  /beer:
    get:
      tags:
      - beer
      summary: Find all beers
      description: Returns beer list
      operationId: getAllBeers
      parameters:
        - name: abv_gt
          in: query
          description: filter by abv greater than assigned value.
          required: false
          schema:
            type: number
            format: double
        - name: abv_lt
          in: query
          description: filter by abv less than assigned value.
          required: false
          schema:
            type: number
            format: double
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeerArray'
            application/xml:
              schema:
                $ref: '#/components/schemas/BeerArray'
    post:
      tags:
      - beer
      summary: Save a new beer to the list
      description: this is a description
      operationId: createBeer
      requestBody:
        description: beer object that needs to be added to the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBeer'
        required: true
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        400:
          description: bad request
      x-codegen-request-body-name: beer
  /beer/{id}:
    get:
      tags:
      - beer
      summary: Find a beers by Id
      description: Returns beer list
      operationId: getBeer
      parameters:
      - name: id
        in: path
        description: ID of beer
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        404:
          description: not found
    put:
      tags:
      - beer
      summary: Save or update a beer by Id
      description: Update description
      operationId: updateBeer
      parameters:
      - name: id
        in: path
        description: ID of beer
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Updated beer object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBeer'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        400:
          description: bad request
      x-codegen-request-body-name: body
    patch:
      tags:
      - beer
      summary: Save or update a beer by Id
      description: Update description
      operationId: softUpdateBeer
      parameters:
      - name: id
        in: path
        description: ID of beer
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Updated beer object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBeer'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beer'
        400:
          description: bad request
      x-codegen-request-body-name: body
    delete:
      tags:
      - beer
      summary: Delete a beer
      description: Delete description
      operationId: deleteBeer
      parameters:
      - name: id
        in: path
        description: ID of beer
        required: true
        schema:
          type: integer
          format: int64
      responses:
        202:
          description: removed element
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeerDelete'
        404:
          description: not found
  /pub:
    get:
      tags:
      - pub
      summary: Find all pubs
      description: Returns pub list
      operationId: getAllPubs
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubArray'
            application/xml:
              schema:
                $ref: '#/components/schemas/PubArray'
    post:
      tags:
      - pub
      summary: Save a new pub to the list
      description: this is a description
      operationId: createPub
      requestBody:
        description: pub object that needs to be added to the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPub'
        required: true
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pub'
        400:
          description: bad request
      x-codegen-request-body-name: beer
  /pub/{id}:
    get:
      tags:
      - pub
      summary: Find a pubs by Id
      description: Returns pub list
      operationId: getPub
      parameters:
      - name: id
        in: path
        description: ID of pub
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pub'
        404:
          description: not found
    put:
      tags:
      - pub
      summary: Save or update a pub by Id
      description: Update description
      operationId: updatePub
      parameters:
      - name: id
        in: path
        description: ID of pub
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Updated pub object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPub'
        required: true
      responses:
        202:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pub'
        400:
          description: bad request
      x-codegen-request-body-name: body
    patch:
      tags:
      - pub
      summary: Save or update a pub by Id
      description: Update description
      operationId: softUpdatePub
      parameters:
      - name: id
        in: path
        description: ID of pub
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Updated pub object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPub'
        required: true
      responses:
        202:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pub'
        400:
          description: bad request
      x-codegen-request-body-name: body
    delete:
      tags:
      - pub
      summary: Delete a pub
      description: Delete description
      operationId: deletePub
      parameters:
      - name: id
        in: path
        description: ID of pub
        required: true
        schema:
          type: integer
          format: int64
      responses:
        202:
          description: removed element
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletePub'
        404:
          description: not found
  /purchase:
    get:
      tags:
      - purchase
      summary: Find all purchases
      description: Returns pub list
      operationId: getAllPurchases
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseArray'
            application/xml:
              schema:
                $ref: '#/components/schemas/PurchaseArray'
    post:
      tags:
      - purchase
      summary: Save a new purchase to the list
      description: this is a description
      operationId: createPurchase
      parameters:
        - name: id_beer
          in: query
          description: id of the beer to buy.
          required: true
          schema:
            type: integer
            format: int64
        - name: id_pub
          in: query
          description: id of the pub where you buy.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: purchase object that needs to be added to the list
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPurchase'
        required: true
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        400:
          description: bad request
      x-codegen-request-body-name: purchase
  /purchase/{id}:
    get:
      tags:
      - purchase
      summary: Find a purchases by Id
      description: Returns purchase list
      operationId: getPurchase
      parameters:
      - name: id
        in: path
        description: ID of purchase
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        404:
          description: not found
    put:
      tags:
      - purchase
      summary: Save or update a purchase by Id
      description: Update description
      operationId: updatePurchase
      parameters:
      - name: id
        in: path
        description: ID of purchase
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Updated purchase object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPurchase'
        required: true
      responses:
        202:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        400:
          description: bad request
      x-codegen-request-body-name: body
    patch:
      tags:
      - purchase
      summary: Save or update a purchase by Id
      description: Update description
      operationId: softUpdatePurchase
      parameters:
      - name: id
        in: path
        description: ID of pub
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: Updated purchase object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPurchase'
        required: true
      responses:
        202:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        400:
          description: bad request
      x-codegen-request-body-name: body
    delete:
      tags:
      - purchase
      summary: Delete a purchase
      description: Delete description
      operationId: deletePurchase
      parameters:
      - name: id
        in: path
        description: ID of purchase
        required: true
        schema:
          type: integer
          format: int64
      responses:
        202:
          description: removed element
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletePurchase'
        404:
          description: not found
components:
  schemas:
    NewBeer:
      type: object
      properties:
        name:
          type: string
          example: Buzz
        tagline:
          type: string
          example: A Real Bitter Experience.
        first_brewed:
          type: string
          example: 09/2007
        description:
          type: string
          example: A light, crisp and bitter IPA brewed with English and American hops.
        image_url:
          type: string
          example: https://images.punkapi.com/v2/keg.png
        abv:
          type: number
          format: double
          example: 4.5
    Beer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Buzz
        tagline:
          type: string
          example: A Real Bitter Experience.
        first_brewed:
          type: string
          example: 09/2007
        description:
          type: string
          example: A light, crisp and bitter IPA brewed with English and American hops.
        image_url:
          type: string
          example: https://images.punkapi.com/v2/keg.png
        abv:
          type: number
          format: double
          example: 4.5
    BeerDelete:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: null
        tagline:
          type: string
          example: null
        first_brewed:
          type: string
          example: null
        description:
          type: string
          example: null
        image_url:
          type: string
          example: null
        abv:
          type: number
          format: double
          example: null
    BeerArray:
      type: array
      items:
        $ref: '#/components/schemas/Beer'
    NewPub:
      type: object
      properties:
        name:
          type: string
          example: MacLaren's
        location:
          type: string
          example: EEUU.
        phone:
          type: string
          example: 666-777-888
    Pub:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: MacLaren's
        location:
          type: string
          example: EEUU.
        phone:
          type: string
          example: 666-777-888
    DeletePub:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: null
        location:
          type: string
          example: null
        phone:
          type: string
          example: null
    PubArray:
      type: array
      items:
        $ref: '#/components/schemas/Pub'
    NewPurchase:
      type: object
      properties:
        price:
          type: number
          format: double
          example: 20
        quantity:
          type: number
          format: double
          example: 4
        date:
          type: string
          example: null
    Purchase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        price:
          type: number
          format: double
          example: 20
        quantity:
          type: number
          format: double
          example: 4
        date:
          type: string
          example: "2021-11-16T09:17:00.4479775"
        beer:
          $ref: '#/components/schemas/Beer'
        pub:
          $ref: '#/components/schemas/Pub'
    DeletePurchase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        price:
          type: number
          format: double
          example: null
        quantity:
          type: number
          format: double
          example: null
        date:
          type: string
          example: null
        beer:
          type: string
          example: null
        pub:
          type: string
          example: null
    PurchaseArray:
      type: array
      items:
        $ref: '#/components/schemas/Purchase'